refactor: Split state from log pages
