+++ .prettierrc
@@ -7,5 +7,8 @@
   "singleQuote": false,
   "tabWidth": 2,
   "trailingComma": "es5",
-  "useTabs": false
+  "useTabs": false,
+  "plugins": [
+    "prettier-plugin-organize-imports"
+  ]
 }
